tools/smart_font_ttf2c/pyproject.toml
[project]
name = "smart_font_ttf2c"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "freetype-py>=2.5.1",
]
